+++ sources/draw_map/draw_map.c
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 22:38:48 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/25 10:50:57 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/25 17:45:35 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,6 @@ int	draw_map(t_data *data)
 		draw_line(data, line);
 	raycast(data);
 	mlx_put_image_to_window(data->mlx, data->win, data->img.img, 0, 0);
-	// mlx_destroy_image(data->mlx, data->img.img);
 	return (0);
 }
