+++ sources/map/valid_chars.c
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 17:10:00 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/25 11:56:44 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/25 17:45:16 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,6 +37,7 @@ t_bool	interactor_chars_or_die(t_map *data, char _char)
 	validator++;
 	return (TRUE);
 }
+
 /******************************************************************************/
 /*									INIT									  */
 /******************************************************************************/
