+++ sources/map/verify_exit.c
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/01 15:14:13 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/24 16:39:33 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/25 17:45:03 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
