+++ sources/map/verify_extension.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   verify_extension.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wwallas- <wwallas-@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 13:30:29 by wwallas-          #+#    #+#             */
-/*   Updated: 2022/12/03 22:59:42 by wwallas-         ###   ########.fr       */
+/*   Updated: 2023/01/25 17:44:10 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ t_bool	verify_extension(char *file_name)
 	extension = ft_strrchr(file_name, '.');
 	if (ft_str_eq(extension, ".cub"))
 		return (TRUE);
-	//printf("Invalid Extension\n");
+	printf("Invalid Extension\n");
 	return (FALSE);
 }
