+++ sources/program/start_program.c
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 13:58:56 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/25 10:59:36 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/25 17:42:21 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,6 @@ void	start_game(t_data *data)
 	printf ("			staring game\n good gameplay XD\n");
 	draw_map_or_die(data);
 	mlx_hook(data->win, EXIT, 0, &cleanup_program, data);
-	//mlx_hook(data->win, FOCUS, 1L << 21, &draw_map, data);
 	mlx_hook(data->win, PRESS_KEY, (1L << 0), &filter_keyboard, data);
 	mlx_loop(data->mlx);
 }
-//10000000W00000001
