+++ sources/raycast/raycast.c
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 14:34:17 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/25 17:37:20 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/25 17:40:01 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,49 +30,6 @@ void	ver_line(t_data *data, t_raycast *itens, int x, int color)
 	y2 = itens->drawEnd;
 	while (++y1 <= y2)
 		my_mlx_pixel_put(&data->img, x, y1, color);
-}
-
-void	nao_sei_oq_ta_acontecendo(t_raycast *itens, t_data *data)
-{
-	double	aux;
-
-	aux = itens->map[P_X] - player_p_x(data) + (1 - itens->step[P_X]) / 2;
-	if (itens->side == 0)
-		itens->perpWallDist = aux / itens->rayDir[P_X];
-	else
-	{
-		aux = itens->map[P_Y] - player_p_y(data) + (1 - itens->step[P_Y]) / 2;
-		itens->perpWallDist = aux / itens->rayDir[P_Y];
-	}
-}
-
-void	lance_ray_in_wall(t_raycast *itens, t_data *data)
-{
-	while (TRUE)
-	{
-		if (itens->sideDist[P_X] < itens->sideDist[P_Y])
-		{
-			itens->sideDist[P_X] += itens->deltaDist[P_X];
-			itens->map[P_X] += itens->step[P_X];
-			itens->side = 0;
-		}
-		else
-		{
-			itens->sideDist[P_Y] += itens->deltaDist[P_Y];
-			itens->map[P_Y] += itens->step[P_Y];
-			itens->side = 1;
-		}
-		print_larger_pixel(data, itens->map[P_X] * 5, \
-		itens->map[P_Y] * 5, RGB_RED);
-		if (data->map.map[itens->map[P_Y]][itens->map[P_X]] == '1')
-			return ;
-	}
-}
-
-void	jump_next_square_and_verify_hit_wall(t_raycast *itens, t_data *data)
-{
-	lance_ray_in_wall(itens, data);
-	nao_sei_oq_ta_acontecendo(itens, data);
 }
 
 void	raycast(t_data *data)
